automations/daily_sales_report/requirements.txt
attrs==22.1.0
autopep8==1.7.0
certifi==2022.6.15.2
charset-normalizer==2.1.1
greenlet==1.1.2
idna==3.4
iniconfig==1.1.1
packaging==21.3
playwright==1.25.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pyee==8.1.0
pyparsing==3.0.9
pytest==7.1.3
pytest-base-url==2.0.0
pytest-playwright==0.3.0
python-dotenv==0.21.0
python-slugify==6.1.2
requests==2.28.1
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.12
websockets==10.1
